restructuring sub components to better organize state transfer.
