+++ client/src/components/overview/sub-comps/StyleThumnails.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function StyleThumbnails({ styles }) {
+  return (
+    <div className="style-thumb-container">
+      <h3>Thumbs Container</h3>
+    </div>
+  );
+}
