+++ client/src/components/questions/styles/questions.css
@@ -10,6 +10,29 @@
   justify-content: space-between;
 }
 
+/* GENERAL TAGS */
+
+.qa h2 {
+  display: block;
+  font-size: 1.5em;
+  margin-block-start: 0.83em;
+  margin-block-end: 0.83em;
+  margin-inline-start: 0px;
+  margin-inline-end: 0px;
+  font-weight: bold;
+}
+
+.qa h3 {
+  display: block;
+  font-size: 1.17em;
+  margin-block-start: 1em;
+  margin-block-end: 1em;
+  margin-inline-start: 0px;
+  margin-inline-end: 0px;
+  font-weight: bold;
+  margin: 1rem 0;
+}
+
 /* QUESTIONS & ANSWERS + GENERALS */
 
 .qa.q\&a {
